backend/pyproject.toml
[project]
name = "dimatech_test_project"
version = "0.1.0"
dependencies = [
    "alembic==1.18.1",
    "fastapi[standard]==0.135.1",
    "psycopg[binary]==3.3.2",
    "PyJWT==2.12.1",
    "pydantic-settings==2.12.0",
    "pwdlib[argon2,bcrypt]==0.3.0",
    "SQLAlchemy==2.0.48",
    "uvicorn[standard]==0.44.0",
]

[project.optional-dependencies]

dev = [
    "black==26.1.0", 
    "flake8==7.3.0",
    "pre-commit==4.5.1",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.black]
line-length = 100
target-version = ["py312"]
exclude = '''
/(
    \.venv
  | alembic/versions/.*
)/
'''
